Moved point cloud and tagger to geometry lib.
