+++ core/lua_core.c
@@ -1395,226 +1395,6 @@ static lua_class_method st_methods[] = {
   {NULL, NULL, NULL}
 };
 
-static lua_module_function tagger_funcs[] = {
-  {NULL, NULL, NULL}
-};
-
-static int t_create_tag(lua_State* L)
-{
-  tagger_t* t = lua_to_tagger(L, 1);
-  if (t == NULL)
-    return luaL_error(L, "Method must be invoked with a tagger.");
-  if (!lua_isstring(L, 2))
-    return luaL_error(L, "Argument 1 must be a string.");
-  if (!lua_istable(L, 3) && !lua_is_array(L, 3, LUA_ARRAY_INT))
-    return luaL_error(L, "Argument 2 must be a table or array of integers.");
-  int* tag = NULL;
-  size_t size = 0;
-  if (lua_is_array(L, 3, LUA_ARRAY_INT))
-  {
-    int_array_t* a = lua_to_array(L, 3, LUA_ARRAY_INT);
-    tag = tagger_create_tag(t, lua_tostring(L, 2), a->size);
-    size = a->size;
-    memcpy(tag, a->data, sizeof(int) * size);
-  }
-  else
-  {
-    int_array_t* a = int_array_new();
-    int i = 1;
-    while (true)
-    {
-      lua_rawgeti(L, 2, i);
-      if (lua_isnil(L, -1))
-      {
-        lua_pop(L, 1);
-        break;
-      }
-      else
-      {
-        if (!lua_isinteger(L, -1))
-          luaL_error(L, "Item %d in argument 1 is not an integer.");
-        int_array_append(a, (int)(lua_tointeger(L, -1)));
-        lua_pop(L, 1);
-      }
-      ++i;
-    }
-    tag = tagger_create_tag(t, lua_tostring(L, 2), a->size);
-    size = a->size;
-    memcpy(tag, a->data, sizeof(int) * size);
-    int_array_free(a);
-  }
-  int_array_t* tt = int_array_new_with_data(tag, size);
-  lua_push_array(L, tt, LUA_ARRAY_INT, true);
-  return 1;
-}
-
-static int t_tag(lua_State* L)
-{
-  tagger_t* t = lua_to_tagger(L, 1);
-  if (t == NULL)
-    return luaL_error(L, "Method must be invoked with a tagger.");
-  if (!lua_isstring(L, 2))
-    return luaL_error(L, "Argument must be a string.");
-  size_t size;
-  int* tag = tagger_tag(t, lua_tostring(L, 2), &size);
-  if (tag != NULL)
-  {
-    int_array_t* tt = int_array_new_with_data(tag, size);
-    lua_push_array(L, tt, LUA_ARRAY_INT, true);
-    return 1;
-  }
-  else
-    return 0;
-}
-
-static int t_has_tag(lua_State* L)
-{
-  tagger_t* t = lua_to_tagger(L, 1);
-  if (t == NULL)
-    return luaL_error(L, "Method must be invoked with a tagger.");
-  if (!lua_isstring(L, 2))
-    return luaL_error(L, "Argument must be a string.");
-  lua_pushboolean(L, tagger_has_tag(t, lua_tostring(L, 2)));
-  return 1;
-}
-
-static int t_tostring(lua_State* L)
-{
-  lua_pushstring(L, "tagger");
-  return 1;
-}
-
-static lua_class_method tagger_methods[] = {
-  {"create_tag", t_create_tag, "tagger:create_tag(name, indices) -> Creates and returns a tag with the given name and indices."},
-  {"tag", t_tag, "tagger:tag(name) -> Returns a tag with the given name."},
-  {"has_tag", t_has_tag, "tagger:has_tag(name) -> Returns true if the tagger has a tag with the given name, false otherwise."},
-  {"__tostring", t_tostring, NULL},
-  {NULL, NULL, NULL}
-};
-
-static int pc_new(lua_State* L)
-{
-  int num_args = lua_gettop(L);
-  if ((num_args != 2) && (num_args != 3))
-  {
-    luaL_error(L, "Arguments must be an MPI communicator, a list of points, "
-                  "and (optionally) a number of ghost points.");
-  }
-
-  if (!lua_is_mpi_comm(L, 1) )
-    luaL_error(L, "Argument 1 must be an MPI communicator.");
-  MPI_Comm comm = lua_to_mpi_comm(L, 1);
-
-  point_t* points = NULL;
-  int num_points = 0;
-  bool free_points = false;
-  if (!lua_istable(L, 2) && !lua_is_array(L, 2, LUA_ARRAY_POINT))
-    luaL_error(L, "Argument 2 must be a list or an array of points.");
-  if (lua_istable(L, 2))
-  {
-    point_array_t* points_array = point_array_new();
-    int i = 1;
-    while (true)
-    {
-      lua_rawgeti(L, 2, i);
-      if (lua_isnil(L, -1))
-      {
-        lua_pop(L, 1);
-        break;
-      }
-      else
-      {
-        if (!lua_is_point(L, -1))
-          luaL_error(L, "Item %d in argument 1 is not a point.");
-        point_t* x = lua_to_point(L, -1);
-        point_array_append(points_array, *x);
-        lua_pop(L, 1);
-      }
-      ++i;
-    }
-    points = points_array->data;
-    num_points = (int)(points_array->size);
-    point_array_release_data_and_free(points_array);
-    free_points = true;
-  }
-  else
-  {
-    points = lua_to_array(L, 2, LUA_ARRAY_POINT);
-    num_points = (int)lua_array_size(L, 2);
-  }
-
-  if ((num_args == 3) && !lua_isinteger(L, 3))
-    luaL_error(L, "Argument 3 must be a number of ghost points.");
-  int num_ghosts = (int)lua_tointeger(L, 3);
-  if (num_ghosts < 0)
-    luaL_error(L, "Number of ghost points must be positive.");
-
-  // Create the point cloud.
-  point_cloud_t* cloud = point_cloud_from_points(comm, points, (int)num_points);
-  if (num_ghosts > 0)
-    point_cloud_set_num_ghosts(cloud, num_ghosts);
-
-  // Clean up.
-  if (free_points)
-    polymec_free(points);
-
-  lua_push_point_cloud(L, cloud);
-  return 1;
-}
-
-static int pc_repartition(lua_State* L)
-{
-  luaL_error(L, "can't repartition point clouds just yet!");
-  return 0;
-}
-
-static lua_module_function pc_funcs[] = {
-  {"new", pc_new, "point_cloud.new(comm, points [, num_ghosts]) -> new point cloud."},
-  {"repartition", pc_repartition, "point_cloud.repartition(cloud) -> Repartitions the given point cloud."},
-  {NULL, NULL, NULL}
-};
-
-static int pc_num_points(lua_State* L)
-{
-  point_cloud_t* pc = lua_to_point_cloud(L, 1);
-  lua_pushinteger(L, pc->num_points);
-  return 1;
-}
-
-static int pc_num_ghosts(lua_State* L)
-{
-  point_cloud_t* pc = lua_to_point_cloud(L, 1);
-  lua_pushinteger(L, pc->num_ghosts);
-  return 1;
-}
-
-static int pc_tags(lua_State* L)
-{
-  point_cloud_t* pc = lua_to_point_cloud(L, 1);
-  lua_push_tagger(L, pc->tags);
-  return 1;
-}
-
-static lua_record_field pc_fields[] = {
-  {"num_points", pc_num_points, NULL},
-  {"num_ghosts", pc_num_ghosts, NULL},
-  {"tags", pc_tags, NULL},
-  {NULL, NULL, NULL}
-};
-
-static int pc_tostring(lua_State* L)
-{
-  point_cloud_t* pc = lua_to_point_cloud(L, 1);
-  lua_pushfstring(L, "point cloud (%d points)", pc->num_points);
-  return 1;
-}
-
-static lua_record_metamethod pc_mm[] = {
-  {"__len", pc_num_points},
-  {"__tostring", pc_tostring},
-  {NULL, NULL}
-};
-
 #ifdef _BSD_SOURCE
 static int posix_rn(lua_State* L)
 {
@@ -2023,8 +1803,6 @@ int lua_register_core_modules(lua_State* L)
   lua_register_class(L, "bbox", "A 3D bounding box.", bbox_funcs, bbox_methods, NULL);
   lua_register_class(L, "sp_func", "A function in 3D space.", sp_funcs, sp_methods, NULL);
   lua_register_class(L, "st_func", "A time-dependent function in 3D space.", st_funcs, st_methods, NULL);
-  lua_register_class(L, "tagger", "An object that holds tags.", tagger_funcs, tagger_methods, NULL);
-  lua_register_record_type(L, "point_cloud", "A point cloud in 3D space.", pc_funcs, pc_fields, pc_mm, DTOR(point_cloud_free));
   lua_register_class(L, "rng", "A random number generator.", rng_funcs, rng_methods, NULL);
 
   // Register the options table.
@@ -2188,36 +1966,6 @@ bool lua_is_st_func(lua_State* L, int index)
 st_func_t* lua_to_st_func(lua_State* L, int index)
 {
   return (st_func_t*)lua_to_object(L, index, "st_func");
-}
-
-void lua_push_tagger(lua_State* L, tagger_t* t)
-{
-  lua_push_object(L, "tagger", t);
-}
-
-bool lua_is_tagger(lua_State* L, int index)
-{
-  return lua_is_object(L, index, "tagger");
-}
-
-tagger_t* lua_to_tagger(lua_State* L, int index)
-{
-  return (tagger_t*)lua_to_object(L, index, "tagger");
-}
-
-void lua_push_point_cloud(lua_State* L, point_cloud_t* c)
-{
-  lua_push_record(L, "point_cloud", c);
-}
-
-bool lua_is_point_cloud(lua_State* L, int index)
-{
-  return lua_is_record(L, index, "point_cloud");
-}
-
-point_cloud_t* lua_to_point_cloud(lua_State* L, int index)
-{
-  return (point_cloud_t*)lua_to_record(L, index, "point_cloud");
 }
 
 void lua_push_rng(lua_State* L, rng_t* r)
