+++ core/tests/CMakeLists.txt
@@ -38,8 +38,6 @@ add_polymec_core_test(test_octree test_octree.c)
 add_polymec_core_test(test_hilbert test_hilbert.c)
 add_polymec_core_test(test_adj_graph test_adj_graph.c)
 add_polymec_core_test(test_special_functions test_special_functions.c)
-add_polymec_core_test(test_point_cloud test_point_cloud.c)
-add_polymec_core_test(test_point_cloud_field test_point_cloud_field.c)
 add_polymec_core_test(test_polynomial test_polynomial.c)
 add_polymec_core_test(test_least_squares test_least_squares.c)
 add_polymec_core_test(test_cxx_program test_cxx_program.cpp)
@@ -47,9 +45,7 @@ add_polymec_core_test(test_cxx_program test_cxx_program.cpp)
 add_mpi_polymec_core_test(test_parallel_timers test_timers.c 1 2 3 4)
 add_mpi_polymec_core_test(test_parallel_kd_tree test_kd_tree.c 1 2 3 4)
 add_mpi_polymec_core_test(test_exchanger test_exchanger.c 1 2 3 4)
-add_mpi_polymec_core_test(test_partition_point_cloud test_partition_point_cloud.c ../../geometry/create_point_lattice.c ../../geometry/cubic_lattice.c 1 2 3 4)
 add_mpi_polymec_core_test(test_parallel_sort test_parallel_sort.c 1 2 3 4)
-add_mpi_polymec_core_test(test_repartition_point_cloud test_repartition_point_cloud.c 1 2 3 4)
 
 # Simple driver test with Lua interpreter.
 include(add_driver_test)
