+++ geometry/lua_geometry.h
@@ -12,6 +12,7 @@
 #include "geometry/coord_mapping.h"
 #include "geometry/sd_func.h"
 #include "geometry/sdt_func.h"
+#include "geometry/point_cloud.h"
 #include "geometry/unimesh.h"
 #include "geometry/polymesh.h"
 
@@ -56,6 +57,28 @@ bool lua_is_sdt_func(lua_State* L, int index);
 // on L's stack, or NULL if the item there is not a temporal signed 
 // distance function.
 sdt_func_t* lua_to_sdt_func(lua_State* L, int index);
+
+// Pushes a tagger t onto L's stack.
+void lua_push_tagger(lua_State* L, tagger_t* t);
+
+// Returns true if the item at the given index on L's stack is a tagger,
+// false if not.
+bool lua_is_tagger(lua_State* L, int index);
+
+// Returns the tagger at the given index on L's stack, or NULL 
+// if the item there is not a tagger.
+tagger_t* lua_to_tagger(lua_State* L, int index);
+
+// Pushes a point cloud c onto L's stack.
+void lua_push_point_cloud(lua_State* L, point_cloud_t* c);
+
+// Returns true if the item at the given index on L's stack is a point
+// cloud, false if not.
+bool lua_is_point_cloud(lua_State* L, int index);
+
+// Returns the point cloud at the given index on L's stack, or NULL 
+// if the item there is not a point cloud.
+point_cloud_t* lua_to_point_cloud(lua_State* L, int index);
 
 // Pushes a unimesh m onto L's stack.
 void lua_push_unimesh(lua_State* L, unimesh_t* m);
