+++ geometry/tagger.c
@@ -5,11 +5,11 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "core/tagger.h"
 #include "core/polymec.h"
 #include "core/unordered_map.h"
 #include "core/unordered_set.h"
 #include "core/array.h"
+#include "geometry/tagger.h"
 
 typedef struct
 {
