+++ geometry/tests/CMakeLists.txt
@@ -15,11 +15,15 @@ add_polymec_geometry_test(test_cylinder_sd_func test_cylinder_sd_func.c)
 add_polymec_geometry_test(test_union_sd_func test_union_sd_func.c)
 add_polymec_geometry_test(test_intersection_sd_func test_intersection_sd_func.c)
 add_polymec_geometry_test(test_difference_sd_func test_difference_sd_func.c)
+add_polymec_geometry_test(test_point_cloud test_point_cloud.c)
+add_polymec_geometry_test(test_point_cloud_field test_point_cloud_field.c)
 add_polymec_geometry_test(test_polymesh test_polymesh.c)
 add_polymec_geometry_test(test_polymesh_field test_polymesh_field.c)
 add_polymec_geometry_test(test_graph_from_polymesh_cells test_graph_from_polymesh_cells.c)
 add_polymec_geometry_test(test_unimesh_patch test_unimesh_patch.c)
 
+add_mpi_polymec_geometry_test(test_partition_point_cloud test_partition_point_cloud.c ../../geometry/create_point_lattice.c ../../geometry/cubic_lattice.c 1 2 3 4)
+add_mpi_polymec_geometry_test(test_repartition_point_cloud test_repartition_point_cloud.c 1 2 3 4)
 add_mpi_polymec_geometry_test(test_create_uniform_polymesh test_create_uniform_polymesh.c 1 2 4)
 add_mpi_polymec_geometry_test(test_create_rectilinear_polymesh test_create_rectilinear_polymesh.c 1 2 3 4)
 add_mpi_polymec_geometry_test(test_crop_polymesh test_crop_polymesh.c 1 2 4)
